Add descriptive initial text to Pattern, Password and PIN Lockscreens
In the security views on lockscreen for PIN/Password and Pattern there
should be an instruction telling the user what is happening. This text
is now shown when no other message is being displayed.

This change was introduced to mitigate confusion that might occur when
the PIN view appears (for example), which may look to many people like
a dialer, especially after they just tried to launch the Dialer app
from the lockscreen.

Change-Id: Idc0c11df9c7165992d3c9e91d4b4e9153d50e29f
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Joe Maples <joe@frap129.org>

Conflicts:
	packages/Keyguard/src/com/android/keyguard/KeyguardPasswordView.java