Performance: Propagate B-services to higher adj
Depending on the inactivity of a service, move the B-services to
highest adj 1000. Under memory pressure, these services will
be killed first ahead of cached apps which results in better
concurrency numbers with bg apps. Inactivity time and minumum
no of services to be maintained are configurable as system properties.
Improves concurrency.

Change-Id: Ied6cfcc3d11951f32f18de680b0e3483db8e163e
Signed-off-by: mydongistiny <jaysonedson@gmail.com>