OMS-N: Support tagging resources as OK to overlay [4/12]
This will allow applications to have a resource xml defining what
resources that are safe to overlay by third party overlay packages.
The format of the tag is <overlay type="..." name="..."/> and it will
result in the FLAG_OVELAY being set on the resource entry.

An overlay package with resources that are not tagged as OK
to overlay by the target application, is considered to be a dangerous
overlay.

Idmaps generated for dangerous overlays will be flagged as dangerous in
the idmap header. It is still possible to use both idmap and overlays that
are dangerous, but it might not be advisable.

The intention is to allow dangerous overlays to be used if
they are pre-installed or if the signature of the overlay package
matches the signature of the target package.

Bug: 31052947

AOSP Change-Id: I0d94e1e208311fe16b07b30aa1c3a32ade01d972

Change-Id: I408107744db307560fd5f171b73e1e9b0395478f
Signed-off-by: mydongistiny <jaysonedson@gmail.com>